Requested changes.
1) Manifest.mf change will allow creation of an editor in a package other than com.arm.cmsis.config
2) Parser change allows a client to control whether the dialog appears on a parser error
3) Scanner change provides a performance improvement for a client scanning 100's of header files

Change 1 allows us to create a new UI on top of the data provided by the parser.
Change 2 allows us to provide a headless client which parses a folder of header files and generates a report of parsing errors.
Change 3 provides an improvement which makes little difference when scanning a single header for an editor, but makes a large difference when scanning 100's of files.